+++ packages/common/src/utils/dump/dump.ts
@@ -1,0 +1,3 @@
+import { Serializable } from '@/types';
+
+export const dump = (value: Serializable) => JSON.stringify(value, null, 2);