+++ packages/ui/src/common/index.ts
@@ -1,1 +1,2 @@
 export * from './enums';
+export * from './types';