+++ packages/ui/src/components/index.ts
@@ -1,0 +1,5 @@
+import '../global.css';
+
+export * from './atoms';
+export * from './molecules';
+export * from './organisms';