+++ services/workflows-service/src/auth/assignee-asigned-guard.service.ts
@@ -1,11 +1,15 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Request } from 'express';
 import { WorkflowService } from '@/workflow/workflow.service';
+import { env } from '@/env';
 
 @Injectable()
 export class WorkflowAssigneeGuard implements CanActivate {
   constructor(private service: WorkflowService) {}
   async canActivate(context: ExecutionContext) {
+    console.log('IS DEMO', env.IS_DEMO);
+    if (env.IS_DEMO) return true;
+
     const request = context.switchToHttp().getRequest<Request>();
     const workflowId = request.params.id;
     const requestingUserId = request.user!.id;