+++ services/workflows-service/src/workflow/schemas/zod-schemas.ts
@@ -14,6 +14,26 @@ export const ConfigSchema = z
     completedWhenTasksResolved: z.boolean().optional(),
     workflowLevelResolution: z.boolean().optional(),
     allowMultipleActiveWorkflows: z.boolean().optional(),
+    callbackResult: z
+      .object({
+        transformers: z.array(z.any()),
+        action: z.string().optional(),
+        deliverEvent: z.string().optional(),
+        persistenceStates: z.array(z.string()).optional(),
+      })
+      .optional(),
+    childCallbackResults: z
+      .array(
+        z
+          .object({
+            definitionId: z.string(),
+            transformers: z.array(z.any()),
+            action: z.string().optional(),
+            deliverEvent: z.string().optional(),
+          })
+          .optional(),
+      )
+      .optional(),
   })
   .strict()
   .optional();