+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @ballerine/headless-example
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.14
+  - @ballerine/workflow-browser-sdk@0.4.17
+
+## 0.0.12
+
+### Patch Changes
+
+- New Build
+
+## 0.0.11
+
+### Patch Changes
+
+- Upgrade packages
+- Updated dependencies
+  - @ballerine/common@0.5.13
+  - @ballerine/workflow-browser-sdk@0.4.16
+
+## 0.0.10
+
+### Patch Changes
+
+- Update versions
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.4.15
+  - @ballerine/common@0.5.12
+
+## 0.0.9
+
+### Patch Changes
+
+- Packages update
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.4.14
+  - @ballerine/common@0.5.11
+
+## 0.0.8
+
+### Patch Changes
+
+- Update packages
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.4.13
+  - @ballerine/common@0.5.10
+
+## 0.0.7
+
+### Patch Changes
+
+- New UI package
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.4.12
+  - @ballerine/common@0.5.9
+
+## 0.0.6
+
+### Patch Changes
+
+- Package update
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.4.11
+  - @ballerine/common@0.5.8
+
+## 0.0.5
+
+### Patch Changes
+
+- Update packages
+- Updated dependencies
+  - @ballerine/common@0.5.7
+  - @ballerine/workflow-browser-sdk@0.4.10
+
 ## 0.0.4
 
 ### Patch Changes